test(e2e): Move helpers to support folder
