+++ test/e2e/default-path.spec.js
@@ -10,7 +10,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 const path = require("path");
 const fsExtra = require("fs-extra");
-const { CliRunner, request, fixturesFolder, wait } = require("./utils");
+const { CliRunner, request, fixturesFolder, wait } = require("./support/helpers");
 
 describe("with no path defined", () => {
   const binaryPath = "./starter";
